All the changes for Qt/Embedded, but not the Qt/Embedded-specific files.
[git-p4: depot-paths = "//depot/qt/main/": change = 20418]

+++ src/tools/qdatastream.cpp
@@ -749,7 +749,7 @@ QDataStream &QDataStream::operator<<( float f )
     CHECK_STREAM_PRECOND
     if ( printable ) {				// printable data
 	char buf[32];
-	sprintf( buf, "%g\n", f );
+	sprintf( buf, "%g\n", (double)f );
 	dev->writeBlock( buf, strlen(buf) );
     } else {
 	float g = f;				// fixes float-on-stack problem
